+++ packages/api/src/crud/users/handlers/user-registration-completed.handler.ts
