+++ packages/api/src/module/read/course-progress/course-progress.rest-controller.ts
@@ -2,10 +2,10 @@ import { Controller, Get, NotFoundException, UseGuards } from '@nestjs/common';
 
 import { GetCourseProgressResponse } from '@coderscamp/shared/models/course-progress';
 
-import { JwtAuthGuard } from '@/auth/jwt/jwt-auth.guard';
-import { JwtUserId } from '@/auth/jwt/jwt-user-id.decorator';
+import { JwtAuthGuard } from '../../../crud/auth/jwt/jwt-auth.guard';
+import { JwtUserId } from '../../../crud/auth/jwt/jwt-user-id.decorator';
 import { PrismaService } from '@/prisma/prisma.service';
-import type { UserId } from '@/users/users.types';
+import type { UserId } from '../../../crud/users/users.types';
 
 @UseGuards(JwtAuthGuard)
 @Controller('course-progress')
