+++ packages/api/src/module/shared/commands/generate-learning-materials-url.domain-command.ts
@@ -1,4 +1,4 @@
-import type { UserId } from '@/users/users.types';
+import type { UserId } from '../../../crud/users/users.types';
 
 export type GenerateLearningMaterialsUrl = {
   type: 'GenerateLearningMaterialsUrl';
