+++ packages/api/src/module/shared/events/learning-materials-url-was-generated.domain-event.ts
@@ -1,4 +1,4 @@
-import type { UserId } from '@/users/users.types';
+import type { UserId } from '../../../crud/users/users.types';
 
 export type LearningMaterialsUrlWasGenerated = {
   type: 'LearningMaterialsUrlWasGenerated';
